[ElfSimulator] fixed some compile errors on this stale project

+++ ElfSimulator/src/pt/up/fe/specs/interpretivesimulator/Program.java
@@ -1,25 +1,19 @@
 /**
- *  Copyright 2020 SPeCS.
+ * Copyright 2020 SPeCS.
  * 
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
  * 
- *       http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  * 
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *  under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License. under the License.
  */
 
 package pt.up.fe.specs.interpretivesimulator;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import pt.up.fe.specs.binarytranslation.instruction.Instruction;
@@ -29,23 +23,22 @@ public class Program {
     private final Map<Integer, Instruction> instructions = new HashMap<>();
     private final int startingInstructionAddress;
     private final int instructionWidth;
-    
-    
+
     public Program(AStaticInstructionStream stream) {
-        
+
         Instruction firstInstruction = stream.nextInstruction();
-        
-        this.startingInstructionAddress = (int) firstInstruction.getAddress();
+
+        this.startingInstructionAddress = firstInstruction.getAddress().intValue();
         this.instructions.put(this.startingInstructionAddress, firstInstruction);
-        
+
         this.instructionWidth = stream.getInstructionWidth();
 
         Instruction instruction = null;
-        while((instruction = stream.nextInstruction()) != null) {
-            this.instructions.put(instruction.getAddress(), instruction);
+        while ((instruction = stream.nextInstruction()) != null) {
+            this.instructions.put(instruction.getAddress().intValue(), instruction);
         }
     }
-    
+
     public Instruction getByAddress(int programCounter) {
         return instructions.get(programCounter);
     }
@@ -57,8 +50,7 @@ public class Program {
     public int getStartingInstructionAddress() {
         return startingInstructionAddress;
     }
-    
-    
+
 }
 
 /*public class Program {
